Initial Centos support

+++ tools/install.sh
@@ -1,5 +1,10 @@
 if [ ! -d /etc/puppet ]; then
-  apt-get install -y git
+  if [ -f /etc/debian_version ]; then
+    apt-get install -y git
+  elif [ -f /etc/redhat-release ]; then
+    yum -y install git
+  fi
+
   git clone https://github.com/cirope/puppet.git /etc/puppet
 
   sh /etc/puppet/tools/install_puppet.sh
